// src/types/DaoConditionalHelper.d.ts
interface DaoConditionalHelperInterface extends ethers.utils.Interface {
  functions: {
    "isAfter(uint256)": FunctionFragment;
    "isDaoMember(address,address)": FunctionFragment;
    "isNotDaoMember(address,address)": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "isAfter",
    values: [BigNumberish]
  ): string;
  encodeFunctionData(
    functionFragment: "isDaoMember",
    values: [string, string]
  ): string;
  encodeFunctionData(
    functionFragment: "isNotDaoMember",
    values: [string, string]
  ): string;

  decodeFunctionResult(functionFragment: "isAfter", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "isDaoMember",
    data: BytesLike
  ): Result;
  decodeFunctionResult(
    functionFragment: "isNotDaoMember",
    data: BytesLike
  ): Result;

  events: {};
}

export class DaoConditionalHelper extends Contract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  on(event: EventFilter | string, listener: Listener): this;
  once(event: EventFilter | string, listener: Listener): this;
  addListener(eventName: EventFilter | string, listener: Listener): this;
  removeAllListeners(eventName: EventFilter | string): this;
  removeListener(eventName: any, listener: Listener): this;

  interface: DaoConditionalHelperInterface;

  functions: {
    isAfter(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;

    "isAfter(uint256)"(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;

    isDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;

    "isDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;

    isNotDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;

    "isNotDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<{
      0: void;
    }>;
  };

  isAfter(timestamp: BigNumberish, overrides?: CallOverrides): Promise<void>;

  "isAfter(uint256)"(
    timestamp: BigNumberish,
    overrides?: CallOverrides
  ): Promise<void>;

  isDaoMember(
    user: string,
    dao: string,
    overrides?: CallOverrides
  ): Promise<void>;

  "isDaoMember(address,address)"(
    user: string,
    dao: string,
    overrides?: CallOverrides
  ): Promise<void>;

  isNotDaoMember(
    user: string,
    dao: string,
    overrides?: CallOverrides
  ): Promise<void>;

  "isNotDaoMember(address,address)"(
    user: string,
    dao: string,
    overrides?: CallOverrides
  ): Promise<void>;

  callStatic: {
    isAfter(timestamp: BigNumberish, overrides?: CallOverrides): Promise<void>;

    "isAfter(uint256)"(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<void>;

    isDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<void>;

    "isDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<void>;

    isNotDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<void>;

    "isNotDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<void>;
  };

  filters: {};

  estimateGas: {
    isAfter(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    "isAfter(uint256)"(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    isDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    "isDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    isNotDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    "isNotDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    isAfter(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    "isAfter(uint256)"(
      timestamp: BigNumberish,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    isDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    "isDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    isNotDaoMember(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    "isNotDaoMember(address,address)"(
      user: string,
      dao: string,
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}